#11503: Move matmul pybind details to cpp file
